feat: move `Enum`s to `enums`
