+++ src/argilla/server/search_engine.py
@@ -20,11 +20,10 @@ from opensearchpy import AsyncOpenSearch, helpers
 from pydantic import BaseModel
 from pydantic.utils import GetterDict
 
-from argilla.server.enums import ResponseStatusFilter
+from argilla.server.enums import FieldType, ResponseStatusFilter
 from argilla.server.models import (
     Dataset,
     Field,
-    FieldType,
     Question,
     QuestionType,
     Record,
